add validations in dto and controllers

+++ ewm-service/src/main/java/ru/practicum/ewm/location/dto/LocationDto.java
@@ -7,8 +7,8 @@ import lombok.Data;
 @AllArgsConstructor
 public class LocationDto {
 
-    private double lat;
+    private float lat;
 
-    private double lon;
+    private float lon;
 
 }
