+++ ewm-service/src/main/java/ru/practicum/ewm/location/repository/LocationRepository.java
@@ -4,5 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import ru.practicum.ewm.location.model.Location;
 
 public interface LocationRepository extends JpaRepository<Location, Long> {
-    Location getLocationByLatAndLon(double lat, double lon);
+
+    Location getLocationByLatAndLon(float lat, float lon);
+
 }
